Custom LU Solver

+++ src/Stencil/stencil.cpp
@@ -5,7 +5,7 @@ Stencil::Stencil(std::initializer_list<int> init)
 {
     std::copy(init.begin(), init.end(), values_.begin());
     stencil_size_ = 0;
-    for (int i = 0; i < init.size(); i++) {
+    for (int i = 0; i < static_cast<int>(init.size()); i++) {
         if (values_[i] != -1)
             stencil_size_++;
     }
